fix: convert risk grades mapping to float before adding base_rate
Add explicit .astype(float) conversion after mapping risk grades to rate premiums to avoid TypeError when adding base_rate to Categorical type. This resolves the "unsupported operand type(s) for +: 'Categorical' and 'float'" error in generate_risk_scores function.

sync: e09672e4

+++ scripts/train.py
@@ -499,7 +499,7 @@ def generate_risk_scores(df, model, selected_features, logger=None):
         'E': 0.10,
         'F': 0.15
     }
-    interest_rates = risk_grades.map(rate_premiums) + base_rate
+    interest_rates = risk_grades.map(rate_premiums).astype(float) + base_rate
     
     # Create output DataFrame
     output_df = pd.DataFrame({
